Fix some cppcheck notes

+++ console/readdvd/imgwriter.cpp
@@ -61,9 +61,11 @@ int imgwriter::write(int lba, int scnt, int ssz, void* buff) {
 		if (fseek(iso, offs, SEEK_SET))
 #endif
 		{
-		    printf("\nseek() failed! Offs: %lld (%08LX)\n", offs, offs);
+			printf("\nseek() failed! Offs: %lld (%08LX)\n",
+				static_cast<long long>(offs),
+				static_cast<unsigned long long>(offs));
 			mutex->unlock();
-		    return 0;
+			return 0;
 		}
 		res = fwrite(buff, ssz, scnt, iso);
 //		printf("\nwrote: %ld of %ld\n", res, scnt);
